glslang spvIR.h: Portability improvement.
git-svn-id: https://cvs.khronos.org/svn/repos/ogl/trunk/ecosystem/public/sdk/tools/glslang@31139 e7fa87d3-cd2b-0410-9028-fcbf551c1848

+++ SPIRV/spvIR.h
@@ -111,7 +111,7 @@ public:
         }
     }
     Op getOpCode() const { return opCode; }
-    int getNumOperands() const { return operands.size(); }
+    int getNumOperands() const { return (int)operands.size(); }
     Id getResultId() const { return resultId; }
     Id getTypeId() const { return typeId; }
     Id getIdOperand(int op) const { return operands[op]; }
@@ -127,9 +127,9 @@ public:
             ++wordCount;
         if (resultId)
             ++wordCount;
-        wordCount += operands.size();
+        wordCount += (unsigned int)operands.size();
         if (string)
-            wordCount += string->size();
+            wordCount += (unsigned int)string->size();
 
         // Write out the beginning of the instruction
         out.push_back(((wordCount) << WordCountShift) | opCode);
